server/requirements.txt
APScheduler==3.7.0
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
Flask==1.1.2
Flask-APScheduler==1.12.2
Flask-SQLAlchemy==2.4.4
idna==3.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-dateutil==2.8.2
python-dotenv==0.15.0
pytz==2021.1
requests==2.26.0
six==1.16.0
soupsieve==2.2
SQLAlchemy==1.3.23
tzlocal==2.1
urllib3==1.26.6
Werkzeug==1.0.1
